+++ packages/store-engine/tsconfig.json
@@ -3,6 +3,6 @@
     "outDir": "dist/commonjs",
     "rootDir": "src/main"
   },
-  "exclude": ["dist", "node_modules"],
+  "exclude": ["dist", "karma.conf.ts", "node_modules"],
   "extends": "../../tsconfig.json"
 }
